requirements.txt
mss==10.1.0
numpy==2.4.4
opencv_python==4.13.0.92
Pillow==12.1.1
psutil==7.2.2
pywin32==311
